refactor: reorganize editor integration tests
