+++ packages/editor/src/__tests__/editor-integration.browser.spec.tsx
@@ -1,12 +1,23 @@
 import { page, userEvent } from 'vitest/browser';
 import { render } from 'vitest-browser-react';
-import { EmailEditor } from './email-editor';
+import { EmailEditor } from '../email-editor/email-editor';
+import { pasteHtml, pasteText } from './browser-test-helpers';
+
+const htmlTemplate = `
+<!doctype html>
+<html>
+  <body>
+    <h1>Pasted heading</h1>
+    <p>Pasted body content</p>
+  </body>
+</html>
+`;
 
 function getEditor() {
   return page.getByRole('textbox');
 }
 
-describe('EmailEditor (browser)', () => {
+describe('editor integration (browser)', () => {
   it('loads and content is editable', async () => {
     render(<EmailEditor />);
 
@@ -115,5 +126,69 @@ describe('EmailEditor (browser)', () => {
     await userEvent.keyboard(`{${mod}>}u{/${mod}}`);
 
     expect(editorEl.innerHTML).toMatch(/<u[ >]/i);
+  });
+
+  it('pasting plain text into an empty editor inserts text', async () => {
+    render(<EmailEditor />);
+
+    const editor = getEditor();
+    await editor.click();
+
+    const editorEl = editor.element() as HTMLElement;
+    pasteText(editorEl, 'hello world');
+
+    await expect.element(editor).toHaveTextContent('hello world');
+  });
+
+  it('pasting plain text into a non-empty editor appends text', async () => {
+    render(<EmailEditor />);
+
+    const editor = getEditor();
+    await editor.click();
+    await userEvent.keyboard('existing');
+
+    const editorEl = editor.element() as HTMLElement;
+    pasteText(editorEl, ' plus pasted');
+
+    await expect.element(editor).toHaveTextContent('existing plus pasted');
+  });
+
+  it('pasting HTML into a non-empty document preserves existing content', async () => {
+    render(
+      <EmailEditor
+        content={{
+          type: 'doc',
+          content: [
+            {
+              type: 'heading',
+              attrs: { level: 1 },
+              content: [{ type: 'text', text: 'Existing heading' }],
+            },
+            {
+              type: 'paragraph',
+              content: [{ type: 'text', text: 'Existing paragraph' }],
+            },
+          ],
+        }}
+      />,
+    );
+
+    const editor = getEditor();
+
+    // Click on the paragraph text to place cursor there, then move to end
+    const para = page.getByText('Existing paragraph');
+    await para.click();
+    await userEvent.keyboard('{End}');
+
+    const editorEl = editor.element() as HTMLElement;
+    pasteHtml(editorEl, htmlTemplate);
+
+    // The original content should still be present
+    expect(editorEl.textContent).toContain('Existing heading');
+    expect(editorEl.textContent).toContain('Existing paragraph');
+
+    // The pasted content should also appear
+    expect(editorEl.textContent).toContain('Pasted heading');
+    expect(editorEl.textContent).toContain('Pasted body content');
   });
 });
